moves columns to their own folder
