+++ app/javascript/mastodon/components/column/back_button.tsx
@@ -7,9 +7,8 @@ import { Icon } from 'mastodon/components/icon';
 import { getColumnSkipLinkId } from 'mastodon/features/ui/components/skip_links';
 import { ButtonInTabsBar } from 'mastodon/features/ui/util/columns_context';
 
-import { useColumnIndexContext } from '../features/ui/components/columns_area';
-
-import { useAppHistory } from './router';
+import { useColumnIndexContext } from '../../features/ui/components/columns_area';
+import { useAppHistory } from '../router';
 
 type OnClickCallback = () => void;
