+++ app/javascript/mastodon/components/column/index.tsx
