+++ app/javascript/mastodon/components/column/search_header.tsx
