+++ app/javascript/mastodon/features/keyboard_shortcuts/index.jsx
@@ -7,8 +7,8 @@ import { Helmet } from '@unhead/react/helmet';
 import ImmutablePureComponent from 'react-immutable-pure-component';
 
 import InfoIcon from '@/material-icons/400-24px/info.svg?react';
-import { Column } from 'mastodon/components/column';
-import { ColumnHeader } from 'mastodon/components/column_header';
+import { Column } from '@/mastodon/components/column';
+import { ColumnHeader } from '@/mastodon/components/column/header';
 import { injectIntl } from '@/mastodon/components/intl';
 
 const messages = defineMessages({
