+++ app/javascript/mastodon/features/terms_of_service/index.tsx
@@ -11,10 +11,10 @@ import { Link, useParams } from 'react-router-dom';
 
 import { Helmet } from '@unhead/react/helmet';
 
+import { Column } from '@/mastodon/components/column';
 import { NavigationFocusTarget } from '@/mastodon/components/navigation_focus_target';
 import { apiGetTermsOfService } from 'mastodon/api/instance';
 import type { ApiTermsOfServiceJSON } from 'mastodon/api_types/instance';
-import { Column } from 'mastodon/components/column';
 import { BundleColumnError } from 'mastodon/features/ui/components/bundle_column_error';
 
 import { getColumnSkipLinkId } from '../ui/components/skip_links';
