+++ app/javascript/mastodon/features/ui/components/column.jsx
@@ -3,7 +3,8 @@ import { PureComponent } from 'react';
 
 import { debounce } from 'lodash';
 
-import ColumnHeader from '../../../components/column_header';
+import { ColumnHeader } from '@/mastodon/components/column/header';
+
 import { isMobile } from '../../../is_mobile';
 import { scrollTop } from '../../../scroll';
