+++ app/javascript/mastodon/features/ui/components/column_loading.tsx
@@ -1,6 +1,6 @@
-import { Column } from 'mastodon/components/column';
-import { ColumnHeader } from 'mastodon/components/column_header';
-import type { Props as ColumnHeaderProps } from 'mastodon/components/column_header';
+import { Column } from '@/mastodon/components/column';
+import { ColumnHeader } from '@/mastodon/components/column/header';
+import type { Props as ColumnHeaderProps } from '@/mastodon/components/column/header';
 
 export const ColumnLoading: React.FC<ColumnHeaderProps> = (otherProps) => (
   <Column>
